% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_a_name.R
\name{draw_a_name}
\alias{draw_a_name}
\title{view the evolution in time of the use of a name for a sex}
\usage{
draw_a_name(the_name, the_sex)
}
\arguments{
\item{the_name}{name}

\item{the_sex}{sex}
}
\value{
a plot
}
\description{
view the evolution in time of the use of a name for a sex
}
\examples{
\dontrun{
draw_a_name("Mathilde","F")
}
}
