% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_a_name.R
\name{draw_names}
\alias{draw_names}
\title{compare the evolution in time of the use of different names}
\usage{
draw_names(the_names)
}
\arguments{
\item{the_names}{names}
}
\value{
a plot
}
\description{
compare the evolution in time of the use of different names
}
\examples{
\dontrun{
draw_a_name(c("Mathilde","Camille"))
}
}
