% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_excel.R
\name{multi_excel}
\alias{multi_excel}
\title{import all sheets from an excel file}
\usage{
multi_excel(file)
}
\arguments{
\item{file}{path to xlsx file}
}
\value{
a list
}
\description{
import all sheets from an excel file
}
\examples{
\dontrun{
multi_excel("/folder/toxlsxfile")
}
}
