% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prenoms_doc.R
\docType{data}
\name{prenoms}
\alias{prenoms}
\title{My dataset}
\description{
My dataset
}
\references{
\url{google.com}{a link}
}
\author{
Mathilde Duverger \email{mathilde1231@hotmail.fr}
}
\keyword{data}
