% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{saves a file as csv}
\usage{
save_as_csv(dataset, filename, row.names = FALSE)
}
\arguments{
\item{dataset}{the data}

\item{filename}{pre existing file}

\item{row.names}{set as FALSE}
}
\description{
saves a file as csv
}
\examples{
\dontrun{
save_as_csv(iris, "output.csv")
}
}
