% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_a_name.R
\name{draw_a_name}
\alias{draw_a_name}
\title{plot the number of people born throughout the years of certain sex and name}
\usage{
draw_a_name(the_name, the_sex)
}
\arguments{
\item{the_name}{name}

\item{the_sex}{sex}
}
\value{
a plot
}
\description{
plot the number of people born throughout the years of certain sex and name
}
\examples{
\dontrun{
draw_a_name("Mathilde","F")
}
}
